+++ src/audio/mod.rs
@@ -12,4 +12,8 @@ pub use interface::{Block, Interface};
 pub const BLOCK_LENGTH: usize = 32; // 32 samples
 pub const HALF_DMA_BUFFER_LENGTH: usize = BLOCK_LENGTH * 2; // 2 channels
 pub const DMA_BUFFER_LENGTH: usize = HALF_DMA_BUFFER_LENGTH * 2; // 2 half-blocks
+
+#[cfg(not(feature = "sampling_rate_96khz"))]
 pub const FS: time::Hertz = time::Hertz::from_raw(48_000);
+#[cfg(feature = "sampling_rate_96khz")]
+pub const FS: time::Hertz = time::Hertz::from_raw(96_000);